requirements.txt
dash==1.21.0
paho-mqtt==1.5.1
plotly==5.1.0
bleak==0.11.0
pandas==1.3.5